+++ itk_dev_shared_components/cvr_api/__init__.py
@@ -1,1 +1,0 @@
-"""For access to the API see https://datacvr.virk.dk/artikel/system-til-system-adgang-til-cvr-data"""